+++ cpp/src/sort/insertion_sort.h
@@ -1,15 +1,14 @@
 #ifndef INSERTION_SORT_H
 #define INSERTION_SORT_H
 
-
-
 #include "utils.h"
+#include<stdio.h>
 
 namespace SORT{
     namespace INSERTION_SORT{
 
     template<typename T>
-    void insertionSortKernel(T arr[], long n);
+    void insertionSortKernel(T arr[], size_t n);
 
     }
 }
