+++ python/AlgoLibR/sort/csort.pxd
@@ -5,16 +5,16 @@ cdef extern from "sort/radix_sort.h" namespace "SORT::RADIX_SORT":
 
 
 cdef extern from "sort/bubble_sort.h" namespace "SORT::BUBBLE_SORT":
-    cdef void bubbleSortKernel[T](T *arr, int n)
+    cdef void bubbleSortKernel[T](T *arr, size_t n)
 
 cdef extern from "sort/quick_sort.h" namespace "SORT::QUICK_SORT":
-    cdef void quickSortKernel[T](T *arr, long n)
+    cdef void quickSortKernel[T](T *arr, size_t n)
 
 cdef extern from "sort/selection_sort.h" namespace "SORT::SELECTION_SORT":
-    cdef void selectionSortKernel[T](T *arr, unsigned int n)
+    cdef void selectionSortKernel[T](T *arr, size_t n)
 
 cdef extern from "sort/insertion_sort.h" namespace "SORT::INSERTION_SORT":
-    cdef void insertionSortKernel[T](T *arr, long n)
+    cdef void insertionSortKernel[T](T *arr, size_t n)
 
 ctypedef fused real:
     cython.short
